cleanup and additions to the README
