+++ README.md
@@ -104,3 +104,56 @@ be loadable in python3 vivisect.
 
 [![CircleCI](https://circleci.com/gh/vivisect/vivisect/tree/master.svg?style=svg)](https://circleci.com/gh/vivisect/vivisect/tree/master)
 [![Build Status](https://travis-ci.org/vivisect/vivisect.svg?branch=master)](https://travis-ci.org/vivisect/vivisect)
+
+## Extending Vivisect / Vdb
+
+Vivisect allows you to extend it's functionality through the use of Vivisect 
+Extensions.  Extensions are loaded with the GUI, and they give nearly complete
+access to the entire Vivisect Workspace and GUI.
+
+Extensions are Python modules loaded from directories contained in the 
+`VIV\_EXT\_PATH` environment variable.  Like DOS or Unix paths, this is a set
+of directories separated by the OS-specific separator (Windows=';', Unix=':').
+
+Like all Python modules, they can be either a `\<modulename\>.py` file or a 
+directory with a `__init__.py` file inside it.  Each module will be loaded into
+the namespace and the `vivExtension(vw, vwgui)` function executed (for Vdb, the
+`vdbExtension(vdb, vdbgui)` function will be executed).  It is up to the module
+to make any modifications (adding menu entries or toolbars, hooking the context
+menu, etc) within this function.  Submodules may be included in the directory-
+version of the extensions, and may be accessed with `from . import \<blah\>`.
+
+In addition to your private zero-day finding extensions, outside plugins may
+be wrapped into Vivisect by simply copying/symlinking them into one of your
+extension directories (listed in the `VIV\_EXT\_PATH`).
+
+If no `VIV\_EXT\_PATH` environment variable has been defined, Vivisect will
+look for extensions in `$HOME/.viv/plugins/` if it exists.  If `VIV\_EXT\_PATH`
+is defined, you much choose to add `$HOME/.viv/plugins/` to it or not.  It will
+not be checked unless it is explicitly listed in `VIV\_EXT\_PATH`.
+
+For examples of using this powerful capability, look at the example file at:
+`vivisect/extensions/example\_gui\_extension.py`
+
+## The Power of Scripts with Vivisect
+
+You can script up meanial tasks or powerful techniques using simple Python
+scripts from either the command-line or the GUI.  
+
+Scripts are loaded and run as any python code is run from the command line.  
+The key diffenece is that Vivisect places a VivWorkspace object in the global
+namespace with the name `vw`.  The GUI, if one exists (Vivisect can be run 
+headless), can be accessed using `vw.getVivGui()`.  
+
+From the CommandLine, analysis modules can be run in the following fashion:
+`$ vivbin -M attackmodule.py targetbin.viv`
+If your module makes any changes to the VivWorkspace, be sure it saves:
+`vw.saveWorkspace()`
+
+To run a script from the GUI, the command bar at the bottom of the screen is
+used. Simply enter:
+`script attackmodule.py \<args\>`
+This method does not need to save to the workspace, as you can choose to do 
+that through standard GUI methods (Ctrl-S or File->Save).
+This method has the added benefit of being able to provide arguments, which
+are placed in the namespace as `argv`.
